login, dashboard and account manager pages code refactored
login, dashboard and account manager pages code refactored
