+++ src/App.js
@@ -1,13 +1,12 @@
-import logo from './logo.svg';
 import './App.css';
-import Login  from "./components/Login";
+import Login  from "./components/login/Login";
 import Dashboard from './components/dashboard/Dashboard';
-import AccountManager from './components/account manager /AccountManager';
+import AccountManager from './components/accountmanager/AccountManager';
 
 function App() {
   return (
         <div className='App'>
-           <AccountManager/>
+           <Dashboard/>
         </div>                            
   ); 
 }
