--- src/components/account manager /AccountManager.css
@@ -1,12 +1,0 @@
-.heading-container{
-    margin-bottom: 65px;
-    margin-right: auto;
-}
-
-
-.dashboard-container{
-    margin: 60px;
-}
-.line{
-    margin : 10px;
-}
