+++ src/components/accountmanager/AccountManager.css
@@ -1,0 +1,13 @@
+.dashboard-container{
+    margin-left: 50px;
+    margin-right: 50px;
+    margin-top: 20px;
+}
+.line{
+    margin : 10px;
+}
+.MuiTableCell{
+    text-align: "Center";
+    font-weight: "bold";
+     font-size:"large";
+}
