+++ src/components/login/Login.css
@@ -5,5 +5,4 @@
 .h2{
     color: #555252;
     margin-top: 0%;
-
 }
